fix(h3-webtransport): reset changes
