+++ crates/h3-webtransport/src/lib.rs
@@ -1,9 +1,9 @@
 //! Provides the client and server support for WebTransport sessions.
 //!
 //! # Relevant Links
-//!
-//! - WebTransport: <https://www.w3.org/TR/webtransport/#biblio-web-transport-http3>
-//! - WebTransport over HTTP/3: <https://datatracker.ietf.org/doc/html/draft-ietf-webtrans-http3/>
+//! WebTransport: <https://www.w3.org/TR/webtransport/#biblio-web-transport-http3>
+//! WebTransport over HTTP/3: <https://datatracker.ietf.org/doc/html/draft-ietf-webtrans-http3/>
+// #![deny(missing_docs)]
 
 /// A WebTransport server
 pub mod server;
